Change logging behaviour
Previously getLogger was called on a module level. This causes the
logger to be (often) created before logging is configured. Changed
logging configuration do not apply to already created loggers (by
default.)

This patch moves the creation of the logger into the functions, which
actually use the logger. This causes the logger to be created after
logging was configured.

Furthermore the logger name was changed to __name__, which will print
the actual module name and the file which caused the log message to be
printed.